+++ src/ray/raylet/node_manager.cc
@@ -1733,7 +1733,7 @@ void NodeManager::HandleRequestWorkerLease(const rpc::RequestWorkerLeaseRequest 
 
   if (new_scheduler_enabled_) {
     auto task_spec = task.GetTaskSpecification();
-    cluster_task_manager_->QueueTask(task, reply);
+    cluster_task_manager_->QueueTask(task, reply, send_reply_callback);
     ScheduleAndDispatch();
     return;
   }
